requirements.txt
Django ==  5.1.4
decouple    == 0.0.7
djangorestframework == 3.15.2
django-cors-headers==4.6.0